Move DependencyModel to molgenis-data util package
